Move tournament matches data to TournamentMatches.h
